BIG UPDATE!
This repository completely broke down, or at least, some files in the .git folder did, and it turned out completely beyond repair.
I did however do a lot of development in the meantime, so I basically just deleted the original .git folder... archived all other files.
Re-cloned the repository so I had a working .git folder
Extracted the archive
And pushed all crap in once...

Sorry, but this was the only way to go with my broken internet and having to wait before I could repair this repository!